New server list for IBM Silverpop:
Silverpop is a subsidiary of IBM that bills itself as "a behavioral marketing platform." For some reason, the company has registered thousands and thousands of domains to support this endeavor. Many of these are used as link targets / landing pages / redirectors in marketing emails. Others, e.g. mkt81.net and pages08.net, have been seen serving analytics and tracking JavaScript to websites.

Because this is such a large chunk of domains, I'm incorporating them as a separate server list that can be toggled on or off in the config file.