Option to set a User-Agent for file_get_contents():
One of the server lists has always taken 3 tries to load successfully; the first
two responses are 0 bytes, then the third contains the list. This is why I added
the retry capability in dad94e8 but it's been continuing to bug me. I'm not sure
what mechanism causes it, but overriding PHP's default empty User-Agent fixes it.