Significant performance improvement:
Refactor the $domains array to be multidimensional, keyed off the first
character of the domain name. This makes all of the calls to in_array()
much faster because each one now only checks a subset of the full list.

This results in vastly improved execution times, at a cost of about 2MB
of extra RAM. One of my servers dropped from 32s to 3s per run average,
and another dropped from 8s to 1s.