When writing an RPZ file, include a meaningful serial number.
In a multi-server primary/secondary scenario, where the nolovia RPZ zone needs
to be transferred to secondary servers from the master, a meaningful serial
number is required, otherwise the master's notify commands will be ignored by
the secondaries.

An incrementing serial number is now included in any generated RPZ file. The
serial is derived from the current time, and is given in the format YYMMDDHHII,
where YY is the last two digits of the current year.