Add a mirrored legacy copy of the defunct hpHosts server list:
This file is a 2020-03-18 backup copy of the hpHosts file, which used to
be located at <http://hosts-file.net/ad_servers.txt>. Malwarebytes has
apparently discontinued maintaining that file. Due to multiple inquiries,
it's been mirrored here, and the default nolovia configuration has been
updated to point at this backup.