Remove s.pinimg.com (fixes #10)
It looks like s.pinimg.com is listed because of some trackers hosted there:

1. hXXps://s.pinimg.com/ct/core.js

2. hXXps://s.pinimg.com/ct/lib/main.04e94784.js

However, having looked at the code a bit, I don't think either of the tracking
scripts will function properly as long as a separate host, ct.pinterest.com, is
still blocked. Which it is, as its own entry, so s.pinimg.com can be removed here.