Filter excessively long hostnames:
Some remote server lists contain FQDNs with excessive lengths. In some cases
these violate RFCs and can cause problems when a name server attempts to load
a zone file. An examination of current data shows that all hosts exceeding 200
characters in length are bogus, so that seems like a reasonable filter to apply.